revamp: merge options page
